Added missing checks for GBT GPU implementation (DAAL) (#1070)
* Added missing checks. Replaced radix sort with common implementation
* Added buffer checks
* Added missing conversion check
* Removed unnecessary conversions
* Replaced uint32_t with unsigned int inside cl kernels

+++ cpp/daal/src/sycl/cl_kernels/radix_sort.cl
@@ -45,41 +45,42 @@ DECLARE_SOURCE(
         //    return x ^ 0x8000000000000000u;
     }
 
-    __kernel void radix_sort_group(__global sortedType * labels, __global int * sorted, __global int * radixbuf, int N, int BlockOffset) {
-        const int global_id = get_global_id(0);
-        const int local_id  = get_local_id(1);
+    __kernel void radix_sort_group(__global sortedType * labels, __global int * sorted, __global int * radixbuf, unsigned int N,
+                                   unsigned int BlockOffset) {
+        const unsigned int global_id = get_global_id(0);
+        const unsigned int local_id  = get_local_id(1);
         // Code is written for a single subgroup. It's necessary to adjust the local range if idle subgoups are presented
         if (get_sub_group_id() > 0) return;
-        const int local_size   = get_sub_group_size();
-        int group_aligned_size = N - N % local_size;
-        int rem                = N - group_aligned_size;
+        const unsigned int local_size   = get_sub_group_size();
+        unsigned int group_aligned_size = N - N % local_size;
+        unsigned int rem                = N - group_aligned_size;
         // radixBuf should be big enough to accumulate radix_range elements
-        const int radix_range = 256;
-        const int byte_range  = 8;
+        const unsigned int radix_range = 256;
+        const unsigned int byte_range  = 8;
 
-        const int radix_count        = sizeof(sortedType);
-        __global sortedType * input  = &labels[global_id * BlockOffset];
-        __global sortedType * output = &sorted[global_id * BlockOffset];
-        __global int * counters      = &radixbuf[global_id * radix_range];
+        const unsigned int radix_count = sizeof(sortedType);
+        __global sortedType * input    = &labels[global_id * BlockOffset];
+        __global sortedType * output   = &sorted[global_id * BlockOffset];
+        __global int * counters        = &radixbuf[global_id * radix_range];
         //  Radix sort
-        for (int i = 0; i < radix_count; i++)
+        for (unsigned int i = 0; i < radix_count; i++)
         {
             __global unsigned char * cinput = (__global unsigned char *)input;
-            for (int j = local_id; j < radix_range; j++) counters[j] = 0;
+            for (unsigned int j = local_id; j < radix_range; j++) counters[j] = 0;
             //  Count elements in sub group to write once per value
-            for (int j = local_id; j < group_aligned_size + local_size; j += local_size)
+            for (unsigned int j = local_id; j < group_aligned_size + local_size; j += local_size)
             {
                 bool exists     = j < group_aligned_size || local_id < rem;
                 unsigned char c = exists ? cinput[j * radix_count + i] : 0;
                 int entry       = -1;
-                for (int k = 0; k < local_size; k++)
+                for (unsigned int k = 0; k < local_size; k++)
                 {
-                    bool correct = j < group_aligned_size || k < rem;
-                    int done     = sub_group_broadcast(correct ? 0 : 1, k);
+                    bool correct      = j < group_aligned_size || k < rem;
+                    unsigned int done = sub_group_broadcast(correct ? 0 : 1, k);
                     if (done) break;
                     unsigned char value = sub_group_broadcast(c, k);
                     if (entry < 0 && value == c) entry = k;
-                    int count = sub_group_reduce_add(exists && value == c ? 1 : 0);
+                    unsigned int count = sub_group_reduce_add(exists && value == c ? 1 : 0);
                     if (entry == local_id && entry == k)
                     {
                         counters[value] += count;
@@ -88,42 +89,42 @@ DECLARE_SOURCE(
                 sub_group_barrier(CLK_GLOBAL_MEM_FENCE);
             }
             //  Parallel scan on counters to generate offsets in place
-            int offset = 0;
-            for (int j = local_id; j < radix_range; j += local_size)
+            unsigned int offset = 0;
+            for (unsigned int j = local_id; j < radix_range; j += local_size)
             {
-                int value          = counters[j];
-                int boundary       = sub_group_scan_exclusive_add(value);
-                counters[j]        = offset + boundary;
-                int partial_offset = sub_group_reduce_add(value);
+                unsigned int value          = counters[j];
+                unsigned int boundary       = sub_group_scan_exclusive_add(value);
+                counters[j]                 = offset + boundary;
+                unsigned int partial_offset = sub_group_reduce_add(value);
                 offset += partial_offset;
             }
             sub_group_barrier(CLK_GLOBAL_MEM_FENCE);
-            for (int j = local_id; j < group_aligned_size + local_size; j += local_size)
+            for (unsigned int j = local_id; j < group_aligned_size + local_size; j += local_size)
             {
-                bool exists      = j < group_aligned_size || local_id < rem;
-                unsigned char c  = exists ? cinput[j * radix_count + i] : 0;
-                int local_offset = 0;
-                int done         = 0;
-                int entry        = -1;
+                bool exists               = j < group_aligned_size || local_id < rem;
+                unsigned char c           = exists ? cinput[j * radix_count + i] : 0;
+                unsigned int local_offset = 0;
+                unsigned int done         = 0;
+                int entry                 = -1;
 
-                for (int k = 0; k < local_size; k++)
+                for (unsigned int k = 0; k < local_size; k++)
                 {
-                    bool correct = j < group_aligned_size || k < rem;
-                    int done     = sub_group_broadcast(correct ? 0 : 1, k);
+                    bool correct      = j < group_aligned_size || k < rem;
+                    unsigned int done = sub_group_broadcast(correct ? 0 : 1, k);
                     if (done)
                     {
                         break;
                     }
-                    int skip = sub_group_broadcast(entry < 0 ? 0 : 1, k);
+                    unsigned int skip = sub_group_broadcast(entry < 0 ? 0 : 1, k);
                     if (skip) continue;
                     unsigned char value = sub_group_broadcast(c, k);
                     if (entry < 0 && value == c) entry = k;
-                    int offset = sub_group_scan_exclusive_add(value == c && exists ? 1 : 0);
+                    unsigned int offset = sub_group_scan_exclusive_add(value == c && exists ? 1 : 0);
                     if (value == c)
                     {
                         local_offset = offset + counters[value];
                     }
-                    int count = sub_group_reduce_add(value == c && exists ? 1 : 0);
+                    unsigned int count = sub_group_reduce_add(value == c && exists ? 1 : 0);
                     if (local_id == k && entry == k)
                     {
                         counters[value] += count;
@@ -134,90 +135,90 @@ DECLARE_SOURCE(
             }
             swap(&input, &output);
         }
-        for (int i = local_id; i < N; i += local_size) output[i] = input[i];
+        for (unsigned int i = local_id; i < N; i += local_size) output[i] = input[i];
     }
 
-    __kernel void radixScan(const __global radixIntType * values, __global int * partialHists, int nRows, int bitOffset) {
-        const int RADIX_BITS = 4;
+    __kernel void radixScan(const __global radixIntType * values, __global int * partialHists, unsigned int nRows, unsigned int bitOffset) {
+        const unsigned int RADIX_BITS = 4;
 
-        const int n_groups             = get_num_groups(0);
-        const int n_sub_groups         = get_num_sub_groups();
-        const int n_total_sub_groups   = n_sub_groups * n_groups;
-        const int nElementsForSubgroup = nRows / n_total_sub_groups + !!(nRows % n_total_sub_groups);
-        const int local_size           = get_sub_group_size();
+        const unsigned int n_groups             = get_num_groups(0);
+        const unsigned int n_sub_groups         = get_num_sub_groups();
+        const unsigned int n_total_sub_groups   = n_sub_groups * n_groups;
+        const unsigned int nElementsForSubgroup = nRows / n_total_sub_groups + !!(nRows % n_total_sub_groups);
+        const unsigned int local_size           = get_sub_group_size();
 
-        const int id           = get_local_id(0);
-        const int local_id     = get_sub_group_local_id();
-        const int sub_group_id = get_sub_group_id();
-        const int group_id     = get_group_id(0) * n_sub_groups + sub_group_id;
+        const unsigned int id           = get_local_id(0);
+        const unsigned int local_id     = get_sub_group_local_id();
+        const unsigned int sub_group_id = get_sub_group_id();
+        const unsigned int group_id     = get_group_id(0) * n_sub_groups + sub_group_id;
 
-        int iStart = group_id * nElementsForSubgroup;
-        int iEnd   = (group_id + 1) * nElementsForSubgroup;
+        unsigned int iStart = group_id * nElementsForSubgroup;
+        unsigned int iEnd   = (group_id + 1) * nElementsForSubgroup;
 
         if (iEnd > nRows)
         {
             iEnd = nRows;
         }
 
-        int offset[1 << RADIX_BITS];
-        const int radix_range   = 1 << RADIX_BITS;
-        const int radix_range_1 = radix_range - 1;
-        for (int i = 0; i < radix_range; i++)
+        unsigned int offset[(unsigned int)1 << RADIX_BITS];
+        const unsigned int radix_range   = (unsigned int)1 << RADIX_BITS;
+        const unsigned int radix_range_1 = radix_range - 1;
+        for (unsigned int i = 0; i < radix_range; i++)
         {
             offset[i] = 0;
         }
 
-        for (int i = iStart + local_id; i < iEnd; i += local_size)
+        for (unsigned int i = iStart + local_id; i < iEnd; i += local_size)
         {
             radixIntType data_bits = ((invBits(values[i]) >> bitOffset) & radix_range_1);
-            for (int j = 0; j < radix_range; j++)
+            for (unsigned int j = 0; j < radix_range; j++)
             {
-                int value          = data_bits == j;
-                int partial_offset = sub_group_reduce_add(value);
+                unsigned int value          = data_bits == j;
+                unsigned int partial_offset = sub_group_reduce_add(value);
                 offset[j] += partial_offset;
             }
         }
 
         if (local_id == 0)
         {
-            for (int j = 0; j < radix_range; j++)
+            for (unsigned int j = 0; j < radix_range; j++)
             {
                 partialHists[group_id * radix_range + j] = offset[j];
             }
         }
     }
 
-    __kernel void radixHistScan(const __global int * partialHists, __global int * partialPrefixHists, int nSubgroupSums) {
-        const int RADIX_BITS = 4;
+    __kernel void radixHistScan(const __global int * partialHists, __global int * partialPrefixHists, unsigned int nSubgroupSums) {
+        const unsigned int RADIX_BITS = 4;
 
         if (get_sub_group_id() > 0) return;
 
-        const int local_size = get_sub_group_size();
-        const int local_id   = get_sub_group_local_id();
+        const unsigned int local_size = get_sub_group_size();
+        const unsigned int local_id   = get_sub_group_local_id();
 
-        int offset[1 << RADIX_BITS];
-        const int radix_range = 1 << RADIX_BITS;
-        for (int i = 0; i < radix_range; i++)
+        unsigned int offset[(unsigned int)1 << RADIX_BITS];
+        const unsigned int radix_range = (unsigned int)1 << RADIX_BITS;
+        for (unsigned int i = 0; i < radix_range; i++)
         {
             offset[i] = 0;
         }
 
-        for (int i = local_id; i < nSubgroupSums; i += local_size)
+        for (unsigned int i = local_id; i < nSubgroupSums; i += local_size)
         {
-            for (int j = 0; j < radix_range; j++)
+            for (unsigned int j = 0; j < radix_range; j++)
             {
-                int value                               = partialHists[i * radix_range + j];
-                int boundary                            = sub_group_scan_exclusive_add(value);
+                unsigned int value                      = partialHists[i * radix_range + j];
+                unsigned int boundary                   = sub_group_scan_exclusive_add(value);
                 partialPrefixHists[i * radix_range + j] = offset[j] + boundary;
-                int partial_offset                      = sub_group_reduce_add(value);
+                unsigned int partial_offset             = sub_group_reduce_add(value);
                 offset[j] += partial_offset;
             }
         }
 
         if (local_id == 0)
         {
-            int totalSum = 0;
-            for (int j = 0; j < radix_range; j++)
+            unsigned int totalSum = 0;
+            for (unsigned int j = 0; j < radix_range; j++)
             {
                 partialPrefixHists[nSubgroupSums * radix_range + j] = totalSum;
                 totalSum += offset[j];
@@ -226,50 +227,50 @@ DECLARE_SOURCE(
     }
 
     __kernel void radixReorder(const __global radixIntType * valuesSrc, const __global int * indicesSrc, const __global int * partialPrefixHists,
-                               __global radixIntType * valuesDst, __global int * indicesDst, int nRows, int bitOffset) {
-        const int RADIX_BITS = 4;
+                               __global radixIntType * valuesDst, __global int * indicesDst, unsigned int nRows, unsigned int bitOffset) {
+        const unsigned int RADIX_BITS = 4;
 
-        const int n_groups             = get_num_groups(0);
-        const int n_sub_groups         = get_num_sub_groups();
-        const int n_total_sub_groups   = n_sub_groups * n_groups;
-        const int nElementsForSubgroup = nRows / n_total_sub_groups + !!(nRows % n_total_sub_groups);
-        const int local_size           = get_sub_group_size();
+        const unsigned int n_groups             = get_num_groups(0);
+        const unsigned int n_sub_groups         = get_num_sub_groups();
+        const unsigned int n_total_sub_groups   = n_sub_groups * n_groups;
+        const unsigned int nElementsForSubgroup = nRows / n_total_sub_groups + !!(nRows % n_total_sub_groups);
+        const unsigned int local_size           = get_sub_group_size();
 
-        const int id           = get_local_id(0);
-        const int local_id     = get_sub_group_local_id();
-        const int sub_group_id = get_sub_group_id();
-        const int group_id     = get_group_id(0) * n_sub_groups + sub_group_id;
+        const unsigned int id           = get_local_id(0);
+        const unsigned int local_id     = get_sub_group_local_id();
+        const unsigned int sub_group_id = get_sub_group_id();
+        const unsigned int group_id     = get_group_id(0) * n_sub_groups + sub_group_id;
 
-        int iStart = group_id * nElementsForSubgroup;
-        int iEnd   = (group_id + 1) * nElementsForSubgroup;
+        unsigned int iStart = group_id * nElementsForSubgroup;
+        unsigned int iEnd   = (group_id + 1) * nElementsForSubgroup;
 
         if (iEnd > nRows)
         {
             iEnd = nRows;
         }
 
-        int offset[1 << RADIX_BITS];
+        unsigned int offset[(unsigned int)1 << RADIX_BITS];
 
-        const int radix_range   = 1 << RADIX_BITS;
-        const int radix_range_1 = radix_range - 1;
+        const unsigned int radix_range   = (unsigned int)1 << RADIX_BITS;
+        const unsigned int radix_range_1 = radix_range - 1;
 
-        for (int i = 0; i < radix_range; i++)
+        for (unsigned int i = 0; i < radix_range; i++)
         {
             offset[i] = partialPrefixHists[group_id * radix_range + i] + partialPrefixHists[n_total_sub_groups * radix_range + i];
         }
 
-        for (int i = iStart + local_id; i < iEnd; i += local_size)
+        for (unsigned int i = iStart + local_id; i < iEnd; i += local_size)
         {
             radixIntType data_value = valuesSrc[i];
             radixIntType data_bits  = ((invBits(data_value) >> bitOffset) & radix_range_1);
-            int pos_new             = 0;
-            for (int j = 0; j < radix_range; j++)
+            unsigned int pos_new    = 0;
+            for (unsigned int j = 0; j < radix_range; j++)
             {
-                int value    = data_bits == j;
-                int boundary = sub_group_scan_exclusive_add(value);
+                unsigned int value    = data_bits == j;
+                unsigned int boundary = sub_group_scan_exclusive_add(value);
                 pos_new |= value * (offset[j] + boundary);
-                int partial_offset = sub_group_reduce_add(value);
-                offset[j]          = offset[j] + partial_offset;
+                unsigned int partial_offset = sub_group_reduce_add(value);
+                offset[j]                   = offset[j] + partial_offset;
             }
             valuesDst[pos_new]  = data_value;
             indicesDst[pos_new] = indicesSrc[i];
